chore(buttons | panels): Minor refactor
