+++ src/main/java/it/units/sdm/quoridor/GUI/buttons/CancelButton.java
@@ -1,9 +1,9 @@
-package it.units.sdm.quoridor.GUI.button;
+package it.units.sdm.quoridor.GUI.buttons;
 
 import it.units.sdm.quoridor.GUI.GUIConstants;
 import it.units.sdm.quoridor.GUI.GameBoardGUI;
 import it.units.sdm.quoridor.GUI.GameController;
-import it.units.sdm.quoridor.GUI.panel.PanelsManager;
+import it.units.sdm.quoridor.GUI.panels.PanelsManager;
 
 import javax.swing.*;
