+++ src/main/java/it/units/sdm/quoridor/GUI/panels/ActionsPanelComponent.java
@@ -1,11 +1,11 @@
-package it.units.sdm.quoridor.GUI.panel;
+package it.units.sdm.quoridor.GUI.panels;
 
 import it.units.sdm.quoridor.GUI.DialogManager;
 import it.units.sdm.quoridor.GUI.GUIConstants;
 import it.units.sdm.quoridor.GUI.GameBoardGUI;
 import it.units.sdm.quoridor.GUI.GameController;
-import it.units.sdm.quoridor.GUI.button.MoveButton;
-import it.units.sdm.quoridor.GUI.button.PlaceWallButton;
+import it.units.sdm.quoridor.GUI.buttons.MoveButton;
+import it.units.sdm.quoridor.GUI.buttons.PlaceWallButton;
 
 import javax.swing.*;
 import java.awt.*;
