+++ src/main/java/it/units/sdm/quoridor/GUI/panels/PanelsManager.java
@@ -1,4 +1,4 @@
-package it.units.sdm.quoridor.GUI.panel;
+package it.units.sdm.quoridor.GUI.panels;
 
 import it.units.sdm.quoridor.GUI.DialogManager;
 import it.units.sdm.quoridor.GUI.GameBoardGUI;
